# api/openapi.yaml
openapi: 3.1.0
info:
  title: XLeaks API
  description: Decentralized P2P social platform API.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local node
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Identity
...
    CreatePostRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 5000
        media_cids:
          type: array
          items:
            type: string
        reply_to:
          type: string
          description: Hex-encoded CID of post being replied to
    PostResponse:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        content:
          type: string
        media_cids:
          type: array
          items:
            type: string
        reply_to:
          type: string
        repost_of:
          type: string
        timestamp:
          type: integer
        like_count:
          type: integer
        tags:
          type: array
          items:
            type: string
...
    CreateReactionRequest:
      type: object
      properties:
        target:
          type: string
          description: Hex-encoded CID of target post
      required: [target]
    ReactionResponse:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        target:
          type: string
        reaction_type:
          type: string
        timestamp:
          type: integer
...
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 500
        website:
          type: string
          maxLength: 200
        avatar_cid:
          type: string
        banner_cid:
          type: string
      required: [display_name]
    ProfileResponse:
      type: object
      properties:
        pubkey:
          type: string
        display_name:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 500
        website:
          type: string
          maxLength: 200
        avatar_cid:
          type: string
        banner_cid:
          type: string
        version:
          type: integer
        timestamp:
          type: integer

    # Follow
    FollowResponse:
      type: object
      properties:
        status:
          type: string
        pubkey:
          type: string
...
    NotificationEntry:
      type: object
      properties:
        type:
          type: string
        actor_pubkey:
          type: string
        actor_display_name:
          type: string
        actor_avatar_cid:
          type: string
        target_post_cid:
          type: string
        related_cid:
          type: string
        timestamp:
          type: integer
        read:
          type: boolean
        target_post_snippet:
          type: string
        reply_snippet:
          type: string
        summary:
          type: string
...
    SearchResponse:
      type: object
      properties:
        type:
          type: string
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
        total:
          type: integer
        message:
          type: string
    SearchHit:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        content:
          type: string
        author:
          type: string
        name:
          type: string
        bio:
          type: string
        score:
          type: number
    TrendingResponse:
      type: object
      properties:
        hashtags:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
              count:
                type: integer
        posts:
          type: array
          items:
            type: object
            properties:
              cid:
                type: string
              author:
                type: string
              content:
                type: string
              like_count:
                type: integer
              reply_count:
                type: integer
              repost_count:
                type: integer
              score:
                type: number
              timestamp:
                type: integer
        window:
          type: string
        message:
          type: string
    ExploreResponse:
      type: object
      properties:
        publishers:
          type: array
          items:
            type: object
            properties:
              pubkey:
                type: string
              follower_count:
                type: integer
              following_count:
                type: integer
              display_name:
                type: string
              bio:
                type: string
        message:
          type: string

    # Media
    MediaUploadResponse:
      type: object
      properties:
        cid:
          type: string
        mime_type:
          type: string
        size:
          type: integer
        chunk_count:
          type: integer
        width:
          type: integer
        height:
          type: integer
        duration:
          type: integer
        thumbnail_cid:
          type: string
        filename:
          type: string
...
  parameters:
...
    PubkeyParam:
      name: pubkey
      in: path
      required: true
      schema:
        type: string
      description: Hex-encoded public key
    BeforeParam:
      name: before
      in: query
      schema:
        type: integer
      description: Cursor timestamp for pagination (milliseconds)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Number of items to return
paths:
...
  /api/posts:
    post:
      tags: [Posts]
      summary: Create a new post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/reactions:
    post:
      tags: [Social]
      summary: Create a reaction (like)
      operationId: createReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReactionRequest'
      responses:
        '201':
          description: Reaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/follow/{pubkey}:
    post:
      tags: [Social]
      summary: Follow a user
      operationId: follow
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
      responses:
        '200':
          description: Followed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResponse'
        '400':
          description: Invalid pubkey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Social]
      summary: Unfollow a user
      operationId: unfollow
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
      responses:
        '200':
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResponse'
        '400':
          description: Invalid pubkey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get own profile
      operationId: getOwnProfile
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Profile]
      summary: Update own profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/search:
    get:
      tags: [Search]
      summary: Search posts or users
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (prefix with # for hashtag search, @ for user search)
        - name: type
          in: query
          schema:
            type: string
            enum: [posts, users]
            default: posts
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Missing query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/trending:
    get:
      tags: [Search]
      summary: Get trending posts and hashtags
      operationId: getTrending
      parameters:
        - name: window
          in: query
          schema:
            type: string
            enum: ['1h', '6h', '24h', '7d']
            default: '24h'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Trending data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendingResponse'
  /api/explore:
    get:
      tags: [Search]
      summary: Discover publishers
      operationId: explore
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Publisher list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExploreResponse'

  # DMs
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Get notifications
      operationId: getNotifications
      parameters:
        - $ref: '#/components/parameters/BeforeParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationEntry'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/media:
    post:
      tags: [Media]
      summary: Upload a media file
      operationId: uploadMedia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '201':
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadResponse'
        '400':
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identity not unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...